+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # CHANGELOG
+## 5.50.0 (February 28, 2024)
+### Components
+#### Added
+* Added: Link web component now supports `mutationObserver` functionality.
+* Added: Marketing icon web component now supports `mutationObserver` functionality.
+* Added: Dropdown web component now supports `retain-selection` attribute.
+* Added: Data table `chiRowClick` event.
+#### Fixed
+* Fixed: Dropdown menu height calculation with header and footer slots present.
+
 ## 5.49.0 (February 21, 2024)
 ### Components
 #### Added